% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barchartandseqlogo.R
\name{Barandseqlogo}
\alias{Barandseqlogo}
\title{A function for generating barchart and seq-logo of co-factors of selected TF}
\usage{
Barandseqlogo(NumberofTop, highestscore, cell, TF, Local = FALSE,
  path = "")
}
\arguments{
\item{NumberofTop}{Number of top co-factors with higher co-binding % to be illustrated}

\item{highestscore}{Co-binding % wich will be the minimum percentage of the shown co-factors.}

\item{cell}{A character string, which is the name of cell under study.}

\item{TF}{A character string which will be the Transcription Factor of interest.}

\item{Local}{A logical value, which will read a local .CSV file in case of TRUE. The file should contain two columns: scores, columnnames which are the co-binding percentages and IDs respectively.}

\item{path}{The path to .CSV file in case Local=TRUE.}
}
\description{
This function generates a barchart of co-binding % for each co-factor of selected TF, along with seq-logo for each of co-factors.
}
