% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaAdder.R
\name{BetaAdder}
\alias{BetaAdder}
\title{A function to merge Beta Score matrices to generate a single matrix.}
\usage{
BetaAdder(TheObject, sp)
}
\arguments{
\item{TheObject}{is an object of S4 class that holds original matrices exported from the package TFregulomeR().}

\item{sp}{is the forking position. User can define up to which position it is required to merge two matrices using this argument.}
}
\value{
This function receives a class object, and returns an updated class object.
}
\description{
This function takes the class object and creates a merge of exclusive Beta Score Matrices
by calculating the elementwise weighted average of them; up to the forking position. Weight of each matrix, is the overlapping percentage of intersectPeakmatrix.
}
\examples{
This function is called within ClassAssignment() function.
}
