% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFPWM.R
\name{createFPWM}
\alias{createFPWM}
\title{A function to generate an FPWM class object.}
\usage{
createFPWM(
  mainTF = NULL,
  partners = NULL,
  cell = NULL,
  mainTF_MMID = NULL,
  partners_MMID = NULL,
  forkPosition = NULL,
  probabilityMatrix = FALSE,
  scaleFrequencyCounts = FALSE,
  flipMatrix = FALSE
)
}
\arguments{
\item{mainTF}{character with the name of the main TF.}

\item{partners}{List or character vector with the names of the partner TFs.}

\item{cell}{character with the name of the main TF.}

\item{mainTF_MMID}{character with the name of the main TF.}

\item{partners_MMID}{character with the name of the main TF.}

\item{forkPosition}{This argument, defines from which point on, the matrix needs to be forked, or in the other words, up to which point two exclusive matrices need to be aggregated.}

\item{probabilityMatrix}{Logical, whether the function should return a frequency matrix or probability matrix (Default FALSE).}

\item{scaleFrequencyCounts}{Logical, whether the count matrix should have equal rowSums across all the rows (Default FALSE).}

\item{flipMatrix}{Logical, whether to apply reverse complement in case the core motif is after the forkPosition (Default FALSE).}
}
\value{
This component, returns a FPWM class object that can be used to plot or written in transfact format.
}
\description{
This function assigns proper data to their associated slots of a S4 classe. This information is either provided by user, or exported from TFregulomeR's dataware using user specified data.
}
\examples{
fpwm <- createFPWM(mainTF ="CEBPB", partners = c("ATF4","ATF7","JUND"), cell = "K562", forkPosition = 5)
}
