% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleobj.R
\name{Ensembles}
\alias{Ensembles}
\title{A function for generating the object for TF in different cell lines}
\usage{
Ensembles(sp, tfname, tfID, CelllinesNumb)
}
\arguments{
\item{sp}{Forking point for final plot}

\item{tfname}{The name of target tf in strings}

\item{tfID}{the targeted motif using MethMotif IDs as the target cell line under study.}

\item{CelllinesNumb}{Maximum number of cell lines to be considered}
}
\description{
This function browses all the cell lines of a given TF and augoments the provided motif with added cell lines to represent the impact of cell line on motif structure
}
