% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembles.R
\name{ensemblesfunc}
\alias{ensemblesfunc}
\title{A function for exporting the motif matrix and augomenting it with additional cell lines}
\usage{
ensemblesfunc(tfname = "JUN", tfID = "MM1_HSA_K562_JUN")
}
\arguments{
\item{tfname}{The name of target tf in strings}

\item{the}{targeted motif using MethMotif IDs as the target cell line under study.}
}
\description{
This function
}
