% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPWMPlotter.R
\name{FPWMPlotter}
\alias{FPWMPlotter}
\title{A function for generating the forked Position Weight Matrix}
\usage{
FPWMPlotter(TheObject, Methylation = TRUE)
}
\arguments{
\item{Methylation}{is a logical value. If it set on TRUE, Methylation level chart will also be plotted. If Flase, only sequence logos will be shown.}

\item{GraphDataObj}{is an object of S4 class with modified and converted data ready to be plotted.}
}
\description{
This function takes the generated class object and plots a forked position weight matrix.
}
