% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixAdder.R
\name{MatrixAdder}
\alias{MatrixAdder}
\title{A function to merge motif matrices to generate one matrix as parent node.}
\usage{
MatrixAdder(TheObject, sp)
}
\arguments{
\item{TheObject}{is a object of S4 class that holds original matrices exported from the package TFregulomeR().}

\item{sp}{is the forking position. User can define up to which position it is required to merge matrices using this argument.}
}
\value{
This function recieves a class object, and returns an updated class object by adding merged matrix to parentmatrix slot
}
\description{
This function takes the object and creates a merge of all matrices 
by claculating the elementwise addition of them, up to a user specified position (Forking Position).
}
\examples{
This function is called whithin ClassAssignment() function. 
}
