% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniCofactorReport.R
\name{miniCofactorReport}
\alias{miniCofactorReport}
\title{Produces a cofactorReport}
\usage{
miniCofactorReport(
  TF,
  cell,
  cobinding_threshold = 0.05,
  Methylation = TRUE,
  includeMotifOnly = TRUE,
  height = 12,
  width = 7,
  pdfName = NULL
)
}
\arguments{
\item{TF}{Main TF of interest.}

\item{cell}{Cell .}

\item{cobinding_threshold}{Only the co-factors with co-binding percentages more than this threshold value will be reported. By default the threshold is 0.05.}

\item{Methylation}{Logical, TRUE to retrieve Cytosine methylation information.}

\item{includeMotifOnly}{Logical, TRUE if you wish to include only peaks that contain the known binding motif.}

\item{height}{Height in inch for the plot.}

\item{width}{Width in inch for the plot.}

\item{pdfName}{Name of the pdf to be saved.}
}
\value{
A PDF file with the cofactorReport.
}
\description{
This function allows you to get a PDF report of top cofactors along with DNA methylation for a TF.
}
\examples{
miniCofactorReport(TF = "CEBPB",cell = "K562")
}
\keyword{cofactorReport}
