% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifyBetaFormat.R
\name{ModifyBetaFormat}
\alias{ModifyBetaFormat}
\title{A function for converting Beta Score matrices into proper data frames.}
\usage{
ModifyBetaFormat(TheObject)
}
\arguments{
\item{TheObject}{is an object of S4 class that holds original matrices exported from the package TFregulomeR().}
}
\value{
This function receives a class object, and returns an updated class object by modifying Beta Score Matrices.
}
\description{
This function receives the S4 class object and converts Betalevel matrices into data frames for better plotting and browsing purposes.
}
\examples{
This function is called within ClassAssignment() function.
}
