% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectGenerator.R
\name{ObjectGenerator}
\alias{ObjectGenerator}
\title{A function togenerate a class object then assign proper data exported from TFregulomeR to its slots.}
\usage{
ObjectGenerator(sp, peak_id_y_list, peak_id_x, height = 2, width = 3)
}
\arguments{
\item{sp}{This argument, defines from which point on, the matrix needs to be forked, or in the other words, up to which point two exclusive matrices need to be aggregated.}

\item{peak_id_y_list}{This argument is a list of TF ID's which will be intersected with Peak_id_x.}

\item{peak_id_x}{This argument holds an id of TFBS compatible with TFregulomeR(). This is the target peak ID wich will be employed by IntersecPeakMatrix of TFregulomeR to extract desired data.}

\item{height}{An argument which allows user to customize the height of final graph relative to screen.}

\item{width}{An argument which allows user to customize the width of final graph relative to screen.}
}
\value{
This component, returns a class object which holds all the neccessary information for other functuins.
}
\description{
This function assigns proper data to their associated slots of a S4 classe. This information is either provided by user, or exported from TFregulomeR's dataware using user specified data.
}
\examples{

}
