% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFPWM.R
\name{plotFPWM}
\alias{plotFPWM}
\title{A function for generating the forked Position Weight Matrix}
\usage{
plotFPWM(
  FPWM,
  Methylation = TRUE,
  height = 12,
  width = 7,
  textSize = 7,
  scaleBarplot = TRUE,
  pdfName = "plotFPWM.pdf"
)
}
\arguments{
\item{FPWM}{is an object of S4 class with modified and converted data ready to be plotted.}

\item{Methylation}{is a logical value. If it set on TRUE, Methylation level chart will also be plotted. If Flase, only sequence logos will be shown.}

\item{height}{Height (inches) of the pdf output.}

\item{width}{Width (inches) of the pdf output.}

\item{textSize}{Size of the text on the plot default 8.}

\item{scaleBarplot}{Logical value, TRUE to keep the  .}

\item{pdfName}{Name for the output pdf.}
}
\description{
This function takes the generated class object and plots a forked position weight matrix.
}
