% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFTRANSFACFile.R
\name{ReadFTRANSFACFile}
\alias{ReadFTRANSFACFile}
\title{A function for generating a class object from a local file in proper format}
\usage{
ReadFTRANSFACFile(File = "MM1_HSA_K562_CEBPB___4-FTRANSFAC.txt",
  Methylation = FALSE)
}
\arguments{
\item{File}{the directory of .txt file}

\item{Methylation}{a logical argument which indicates if Methylation Score files are provided and needed to be included in Object or not.}
}
\value{
A class object for plotting. The Methylation Score matrices can be optionally ommitted or not.
}
\description{
This function reads an stored .txt file of FTRANSFAC format and constructs a class object from it. As default, the returned class Object does not contain Methylation Score matrices. If needed, files exported from TFregulomeR() with the same name and format should be provided before setting MEthylation==TRUE.
}
