% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoreMultiTRANSFACFile.R
\name{StoreMultiTRANSFACFile}
\alias{StoreMultiTRANSFACFile}
\title{Generating and storing a .txt file named "All.txt" which contains multiple FPWMs, contatinated together, respecting TRANSFAC format.}
\usage{
StoreMultiTRANSFACFile(List_sp, Listof_peak_id_y_list, List_peak_id_x)
}
\arguments{
\item{List_sp}{List of forking position numbers for each one of FPWM.}

\item{Listof_peak_id_y_list}{A list of lists. Each list within this list, is a set of IDs which are going to form one FPWM plot.}

\item{List_peak_id_x}{A list of IDs.  The ID in List_peak_id_x[i] will be employed to form multiple IntersecPeakMatrices with all the IDs exisiting in Listof_peak_id_y_list[i].}
}
\value{
This function stores a .txt file at working directory, and returns name of the file for more convenience.
}
\description{
This function generates a .txt file which holds number of the data structures needed for one set of plotting in FTRANSFAC format.
}
