% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storeTRANSFAC.R
\name{storeTRANSFAC}
\alias{storeTRANSFAC}
\title{A function for storing TRANSFAC files of the forked matrices.}
\usage{
storeTRANSFAC(TheObject)
}
\arguments{
\item{TheObject}{the input is the object of FPWM class that holds the raw matrices directly exported from TFregulomeR().}
}
\description{
This function generates files of regular TRANSFAC format in order to further analysis and evaluation. Each file name holdes the name of Transfactor of interest, and the co-factor that is under analysis in the current matrix.
}
