% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToTFBStools.R
\name{ToTFBSTools}
\alias{ToTFBSTools}
\title{A function for generating object of TFBStools holding PFM of each fokred matrix.}
\usage{
ToTFBSTools(TheObject)
}
\arguments{
\item{TheObject}{is the object of the FPWM class. It needs to contain the matrices, IDs and parent matrix.}
}
\description{
This function generates a TFBStools object of each matrix present in FPWM class object, and returns a list containing all the objects.
}
