% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.FPWM.R
\name{write.FPWM}
\alias{write.FPWM}
\title{Writes the FPWM in transfact format}
\usage{
write.FPWM(FPWM = NULL, format = "transfac", fileName = NULL)
}
\arguments{
\item{FPWM}{FPWM object.}

\item{format}{the "transfac" option will output a standard transfact matrix per binding partner. The "FPWMtransfac" prints a single matrix with all the binding partners in FPWM format.}

\item{fileName}{name of the file where the FPWM is going to be written.}
}
\value{
A PDF file with the cofactorReport.
}
\description{
This function saves the FPWM in transfact format.
}
\examples{
write.FPWM(FPWM = fpwm, format = "transfac", fileName = "FPWM.transfact" )
write.FPWM(FPWM = fpwm, format = "FPWMtransfac", fileName = "FPWM.FPWMtransfac" )
}
\keyword{cofactorReport}
